Fixes <http://webkit.org/b/30792>. Web Inspector: When changing resource scope, resource graph can get in weird state.
Reviewed by Timothy Hatcher.

If the user was scrolled down while looking at All resources, and then just
selects Fonts, or something that doesn't have many resources, then the graph
will stay scrolled down, even when there is no content there.
        
When the filter is updated, scroll to the top of the graph view, so this won't
happen.

* inspector/front-end/AbstractTimelinePanel.js:
(WebInspector.AbstractTimelinePanel.prototype._updateFilter):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk/WebCore/inspector/front-end@50098 268f45cc-cd09-0410-ab3c-d52691b4dbfc